README.md
We know that sometimes jotting something down or making a note of a dream or goal is all it takes to get started. That is why we have designed Journey Journals, so that users are able to be productive in working towards their goals and stepping outside of their comfort zones!
We initially built the frameworks for our website on repl.it, which allowed for our team to collaborate in real time together!   Afterwards, we transferred our code to Microsoft Azure in order to make a static web app with a personalized domain.